requirements.txt
alembic==1.11.2
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
bcrypt==4.0.1
blinker==1.6.2
click==8.1.6
colorama==0.4.6
comm==0.1.4
debugpy==1.6.7.post1
decorator==5.1.1
distlib==0.3.6
Django==4.2.1
dominate==2.8.0
executing==1.2.0
filelock==3.12.2
Flask==2.3.2
Flask-Bootstrap==3.3.7.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
ipykernel==6.25.1
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.1
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mysql-connector==2.2.9
mysql-connector-python==8.1.0
mysqlclient==2.2.0
nest-asyncio==1.5.7
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.8.1
prompt-toolkit==3.0.39
protobuf==4.21.12
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
pywin32==306
pyzmq==25.1.1
reportlab==4.0.4
six==1.16.0
SQLAlchemy==2.0.19
sqlparse==0.4.4
stack-data==0.6.2
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
virtualenv==20.23.1
visitor==0.1.3
wcwidth==0.2.6
Werkzeug==2.3.6
WTForms==3.0.1
